feat(GameServer): 收敛物品处理与服务器运行时恢复
模型/工具来源：gpt-5.6-luna；GameServer.exe IDA MCP port 10004；PDB、IDA 反编译/汇编、直接调用者与当前 CMake 边界交叉核对。

F/_PROGRAM_HG/Source/Soulworker/Common/XNet/XCommon/PSCommon.h
F/_PROGRAM_HG/Source/Soulworker/Common/XNet/XCommon/PSServer/PSServerCashShop.h
F/_PROGRAM_HG/Source/Soulworker/Common/XNet/XCommon/PSServer/PSServerCore.h
F/_PROGRAM_HG/Source/Soulworker/Common/XNet/XCommon/PSServer/PSServerMail.h
F/_PROGRAM_HG/Source/Soulworker/Common/XNet/XIOCPBase/Parse.h
- 物品移动、使用、合成、分割、Tick、冷却、现金里程与相关序列化协议恢复：修正字段布局、非 const 运算符、列表计数和字节 ABI。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/Process/ItemProcess.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/Process/ItemProcess.h
- ReqItemMove (0x140453630) 恢复：补齐客户端物品移动的用户边界、库存转移和后续更新路径。
- ReqItemCombine (0x140454050) 恢复：按 IDA 调用顺序取得 CUser 后解析请求并执行合成路径。
- ReqItemDivide (0x1404556B0) 恢复：补齐分割请求解析、数量处理和物品更新路径。
- ReqItemBreak (0x1404569D0) 恢复：补齐破坏请求的完整非空处理逻辑。
- ReqItemUse (0x140458C20) 恢复：接入物品使用校验、效果处理、消耗和响应路径。
- ResEquipSlotOpen (0x140464FC0) 恢复：保留响应包顺序、区域分支和装备栏更新。
- ResItemUseInfoSelect (0x140465500) 恢复：复制具体 ST_USE_ITEM_INFO_LIST 并同步使用状态。
- ResItemUseInfoUpdate (0x140465860) 恢复：同步日期、tick 与客户端/数据库更新路径。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/actor/component/GocInventory.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/actor/component/GocInventory.h
- IsValidEquipItem (0x1400E0CB0) 修复：恢复 TB_ITEM/TB_ITEM_CLASSIFY 查询、XItemFactory::IsCanEquip 判定和精确错误分支。
- CanEquipSlotOpen (0x1400B6570)、EquipSlotOpen (0x1400B6810)、ChangeEquipSlotPos (0x1400B6D20)、CheckEquipSlotOpen (0x1400B6E90)、EquipSlotOpen(int) (0x1400B7090)、SendEquipSlotOpen (0x1400B70D0)、SetEquipSlot (0x1400B71B0)、EquipSlotOpenCalc (0x1400B71E0) 恢复：还原装备栏位映射、等级门槛、位掩码与更新路线。
- SetItemUseInfoList (0x1400B7370)、OnInitItemUseInfoDate (0x1400B74F0)、CanUseItemInfo (0x1400B76B0) 恢复：还原使用信息列表、日期/tick 更新和可用性判断。
- 删除无 PDB decorated symbol、无直接调用者证明的 CGocInventory::CanEquip 伪实现。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/XItemFactory.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/XItemFactory.h
- IsCanEquip (0x1408B16C0) 恢复：实现装备槽范围 lower_bound、边界检查和闭区间判定。
- 补齐 XItemFactory 的 PDB-backed 构造、初始化、物品序列号、标题、broach 与 random-option 支持。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/GameDBSocket.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/GameDBSocket.h
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/User.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/User.h
- 收敛 DBParse、系统响应、物品使用信息和 world-event 数据边界，保持当前 active CMake target 的调用链。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/XCore/XServer/*.h
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/ThreadLocalData_Stub.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XCore/XArea/*
- 恢复 PDB-backed 运行时对象布局、锁、对象池、客户端基类、XActor/XArea 边界和 active stub 支持；新增 GreenDamTan_StdCompat、GreenDamTan_TxMap、GreenDamTan_TxPool 仅用于保持当前跨平台编译边界和已证实布局。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XControlServer/ControlServer.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XLoginServer/PartyManager.h
- 修正 XItemFactory::Init 的三参数静态调用和 Windows socket 头文件顺序，避免 winsock.h/winsock2.h 重定义。

F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/Process/LoginProcess.*
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/LogicThreadManager_Stub.cpp
F/_PROGRAM_HG/Source/Soulworker/GameServer/XRelayServer/Thread/LogicThreadProcessor.*
F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/actor/component/*
F/_PROGRAM_HG/Source/Soulworker/GameServer/XSCommon/Table/TB_REPACKAGECOSTUME.h
- 同步物品、交换、邮件、Akashic、社交物品、真空组、重封装服装表及线程处理依赖的已恢复 ABI、字段和调用边界。

CLAUDE.md
docs/GameServer.exe-func-index.md
docs/GameServer.exe-type-index.md
docs/GameServer.exe-path-recovery-index.md
docs/GameServer.exe-current-target-progress.md
- 同步当前 GameServer 目标规则、PDB/IDA 函数与类型索引、路径索引和 append-only 进度记录；清理函数/类型重复键，移除未证明的 CanEquip 条目，并保留外部 Common DB connection failure 的非敏感描述。

验证
- 独立 verifier：VERDICT PASS；四目标 clean build、GameServer smoke、IDA/PDB 对照、台账审计、UTF-8、CTest 与 git diff --check 均通过。
- cmake --build build --target LoginServer RelayServer ControlServer GameServer --clean-first -- -j1：四目标 exit 0，分别链接 LoginServer.exe、RelayServer.exe、ControlServer.exe、GameServer.exe。
- cmake --build build --target GameServer --clean-first -- -j1：exit 0；无 error、fatal error、FAILED 或 ninja error，仅既有编译 warning。
- GREENDAMTAN_AUTOSTOP_MS=5000 timeout 45s ./build/bin/GameServer.exe：exit 0，达到 Complete Server Init 和 Auto shutdown tick reached；forbidden smoke pattern count=0。
- CTest：No tests were found!!!。
- 函数索引 58585 行、类型索引 180 行；独立解析 duplicate_rows=0，verified 描述中的 TODO/stub/placeholder/unimplemented=0。

+++ F/_PROGRAM_HG/Source/Soulworker/GameServer/XGameServer/ThreadLocalData.h
@@ -315,7 +315,7 @@ public:
     void SendOperationTime();
 
     // IDA @ 0x1406DA780 - Send world event booster
-    void SendWorldEventBooster(unsigned int dwBuff_ID, __int64 biEndDate);
+    void SendWorldEventBooster(unsigned long dwBuff_ID, __int64 biEndDate);
 
     // IDA @ 0x1406DA820 - Compare league inventory count
     bool CompareLeagueInventoryCount(unsigned int nLeagueID, unsigned int dwUCID);
